Added alert to confirm deletion of best times

+++ app/src/main/java/com/example/android_minesweeper/screens/best_times/HighScoresFragment.kt
@@ -1,6 +1,7 @@
 package com.example.android_minesweeper.screens.best_times
 
 import android.annotation.SuppressLint
+import android.app.AlertDialog
 import android.os.Bundle
 import android.view.*
 import androidx.appcompat.app.AppCompatActivity
@@ -8,6 +9,7 @@ import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProviders
 import com.example.android_minesweeper.R
+import com.example.android_minesweeper.databinding.ConfirmDeleteAlertBinding
 import com.example.android_minesweeper.models.AppDatabase
 import com.example.android_minesweeper.databinding.HighScoresScreenBinding
 import com.example.android_minesweeper.view_models.HighScoresViewModel
@@ -19,16 +21,22 @@ class HighScoresFragment : Fragment() {
     private lateinit var viewModelFactory: HighScoresViewModelFactory
 
     @SuppressLint("RestrictedApi")
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
 
         (activity as? AppCompatActivity)?.supportActionBar?.setShowHideAnimationEnabled(false)
         (activity as? AppCompatActivity)?.supportActionBar?.show()
 
         binding = DataBindingUtil.inflate(inflater, R.layout.high_scores_screen, container, false)
 
-        val dataSource = AppDatabase.getInstance(requireNotNull(this.activity).application).highScoreDao
+        val dataSource =
+            AppDatabase.getInstance(requireNotNull(this.activity).application).highScoreDao
         viewModelFactory = HighScoresViewModelFactory(dataSource)
-        viewModel = ViewModelProviders.of(this, viewModelFactory).get(HighScoresViewModel::class.java)
+        viewModel =
+            ViewModelProviders.of(this, viewModelFactory).get(HighScoresViewModel::class.java)
         binding.highScoresViewModel = viewModel
         binding.lifecycleOwner = this
 
@@ -50,7 +58,27 @@ class HighScoresFragment : Fragment() {
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         if (item.itemId == R.id.delete_best_times_button) {
-            viewModel.deleteBestTimesButtonPressed()
+
+            DataBindingUtil.inflate<ConfirmDeleteAlertBinding>(
+                layoutInflater,
+                R.layout.confirm_delete_alert,
+                view as? ViewGroup,
+                false
+            ).also { binding ->
+                AlertDialog.Builder(context).create().also { dialog ->
+                    dialog.setView(binding.root)
+                    dialog.setCancelable(false)
+                    dialog.show()
+
+                    binding.cancelDeleteButton.setOnClickListener {
+                        dialog.cancel()
+                    }
+                    binding.confirmDeleteButton.setOnClickListener {
+                        dialog.cancel()
+                        viewModel.deleteBestTimesButtonPressed()
+                    }
+                }
+            }
         }
         return super.onOptionsItemSelected(item)
     }
